refactor(routing): migrate basecamp to root and draft to studio
Why:
- As SiteCue transitions to a dedicated app subdomain (app.sitecue.app), the dashboard should serve as the true root (Launchpad).
- To streamline the URL structure and prepare for the upcoming "Weave Studio" and "Note Management" features without breaking existing workflows.

What:
- Promoted `/basecamp` dashboard pages to the root directory `/`.
- Migrated `/basecamp/draft/*` to `/studio/*` to establish the new writing workspace.
- Relocated the legacy notes list from `src/app/page.tsx` to `src/app/notes/page.tsx` to preserve the "Your Cues" functionality.
- Updated internal link paths and `middleware.ts` to reflect and protect the new routes.
- Maintained the legacy `/weave` directory untouched to ensure backward compatibility with the Chrome extension currently under review.
