+++ apps/app/src/app/notes/page.tsx
@@ -1,0 +1,100 @@
+import Link from "next/link";
+import { redirect } from "next/navigation";
+import { createClient } from "@/utils/supabase/server";
+
+export default async function Dashboard() {
+	const supabase = await createClient();
+
+	const {
+		data: { user },
+	} = await supabase.auth.getUser();
+
+	if (!user) {
+		return redirect("/login");
+	}
+
+	const { data: notes } = await supabase
+		.from("sitecue_notes")
+		.select("*")
+		.eq("user_id", user.id)
+		.order("created_at", { ascending: false });
+
+	return (
+		<div className="min-h-screen bg-gray-50 font-sans">
+			<nav className="bg-white border-b border-gray-200">
+				<div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+					<div className="flex justify-between h-16">
+						<div className="flex">
+							<div className="shrink-0 flex items-center">
+								<Link
+									href="/"
+									className="text-xl font-bold hover:text-gray-700"
+								>
+									SiteCue
+								</Link>
+							</div>
+						</div>
+						<div className="flex items-center">
+							<span className="text-sm text-gray-500 mr-4">{user.email}</span>
+							<form
+								action={async () => {
+									"use server";
+									const supabase = await createClient();
+									await supabase.auth.signOut();
+									redirect("/login");
+								}}
+							>
+								<button
+									type="submit"
+									className="text-sm font-medium text-gray-700 hover:text-black"
+								>
+									Sign out
+								</button>
+							</form>
+						</div>
+					</div>
+				</div>
+			</nav>
+
+			<main className="max-w-7xl mx-auto py-6 sm:px-6 lg:px-8">
+				<div className="px-4 py-6 sm:px-0">
+					<div className="flex justify-between items-center mb-6">
+						<h1 className="text-2xl font-semibold text-gray-900">Your Cues</h1>
+					</div>
+
+					{notes && notes.length > 0 ? (
+						<div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-3">
+							{notes.map((note) => (
+								<div
+									key={note.id}
+									className="bg-white overflow-hidden shadow rounded-lg hover:shadow-md transition-shadow"
+								>
+									<div className="px-4 py-5 sm:p-6">
+										<div className="flex items-center justify-between mb-2">
+											<span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-blue-100 text-blue-800 truncate max-w-[200px]">
+												{note.url_pattern}
+											</span>
+											<span className="text-xs text-gray-400">
+												{note.created_at.substring(0, 10).replace(/-/g, "/")}
+											</span>
+										</div>
+										<p className="text-sm text-gray-900 whitespace-pre-wrap">
+											{note.content}
+										</p>
+									</div>
+								</div>
+							))}
+						</div>
+					) : (
+						<div className="text-center py-12 bg-white rounded-lg shadow">
+							<p className="text-gray-500">You haven't created any cues yet.</p>
+							<p className="text-sm text-gray-400 mt-2">
+								Use the Chrome Extension to add cues to websites.
+							</p>
+						</div>
+					)}
+				</div>
+			</main>
+		</div>
+	);
+}
